fix(system-audit): Fix-Welle B — Codex + CodeRabbit-CLI 4-Wege-Review-Findings
Post-Impl-Reconciliation Task 13. TDD RED-GREEN-REFACTOR mit Red-Green-Cycle-Verify
fuer die Kernregression (Finding E stash-revert -> MUST FAIL -> restore -> PASS).

Findings umgesetzt:
  E (Codex major) — rc=2 Info-Loss behoben: main() rendert partial audit auf
    stdout (JSON: partial=True + internal_errors[], human: ueblicher Report
    + stderr-Annotation) BEVOR rc=2 escaliert. STATE.md bleibt ungeschrieben
    (korrumpierter Audit-Stand nicht persistiert). 2 neue render_json-Tests
    + 1 neuer main-integration-test mit synthetischem broken-check.
  A (Codex+CodeRabbit major) — Thread-Leak: Daemon-Thread-Subclass in Py 3.14
    nicht moeglich (daemon=True auf active thread raises RuntimeError).
    Docstring-Klarstellung "SOFT timeout, CLI one-shot, backlog: subprocess-
    based hard timeout fuer Task 14/15".
  B (beide minor) — vault_timeout_s: int=20 -> float=20.0, int(...) cast in
    main() entfernt. Kein stillen Truncation-Gap mehr (--timeout 2.5 -> 2.5
    statt 2, kein Mismatch zu _run_one).
  F (Codex minor) — Timeout-FailureDetail-category: pre-fix hardcoded "core",
    jetzt via ScopedCheck-Tupel (name, spec, category) an _run_one propagiert.
    Correct Labeling fuer Optional-Checks (Task 14).
  C (beide minor) — _scope_label unused entfernt.
  L (Subagent nit) — datetime.timezone.utc -> datetime.UTC (Py 3.11+ Idiom).
  CR-1 (CodeRabbit-CLI potential_issue) — rc-check vor json.loads in
    test_orchestrator_vault_empty_registry_emits_skip (und K: dito in
    test_orchestrator_duration_budget_core). Silent-rc=2-false-pass verhindert.
  CR-2 (CodeRabbit-CLI nit) — _load_run Return-Type-Hint
    Callable[[Path, AuditContext], CheckResult].

Nicht umgesetzt (ohne CLI-Bestaetigung, Subagent-Ersatz-only, lower priority):
  H (sys.path Import-Side-Effect), I (assert-unter-python-O), m3-m8 + Nits.
  Backlog fuer Task 14/15.

Verification:
  - 12/12 smoke suites PASS (fresh run)
  - --core rc=1 4/8 PASS 2 FAIL 2 WARN (stable vs pre-Welle — echte Drift)
  - --minimal-baseline rc=0 3/3 PASS
  - --vault rc=0 expliziter SKIP (no_checks_registered)
  - Red-Green-Verify-Cycle Finding E: stash -> FAIL JSONDecodeError (stdout
    leer, exakt erwartet) -> pop -> PASS (12/12)

Codex Post-Impl-Attestierung (aus Task-13-Review):
  Spec-§-Drift-Reconciliation: PASS (6 Spec-§6.1-Flags unveraendert,
  --minimal-baseline + --timeout-per-check rein additiv, Plan-Header-Notice
  dokumentiert). Release-reif fuer Task-17-Live-Baseline-Run.

+++ 03_Tools/system_audit/types.py
@@ -39,4 +39,4 @@ class AuditContext:
     """
     repo_root: Path
     include_optional: bool = False
-    vault_timeout_s: int = 20
+    vault_timeout_s: float = 20.0
